important changes in model 20-2-2021

+++ DMeServices.Models/Common/UserCom.cs
@@ -15,7 +15,7 @@ namespace DMeServices.Models.Common
     {
         #region Method :: User By Civil Id
 
-        public static User UserByCivilID(long CivilId)
+        public static User UserByCivilID(int CivilId)
         {
             using (eServicesEntities db = new eServicesEntities())
             {
